pyproject.toml
[project]
name = "hallucination-mitigation"
version = "0.1.0"
description = "Demo notebooks for AI hallucination mitigation workshop"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "awslabs-aws-api-mcp-server>=0.3.1",
    "awslabs-aws-pricing-mcp-server>=1.0.12",
    "boto3>=1.39",
    "datasets>=4",
    "ipykernel>=6.30.1",
    "langchain>=1.0.5",
    "langchain-aws>=1",
    "langfuse>=3.5.0",
    "numpy>=1.26.4",
    "openai>=1.109",
    "python-dotenv>=1.1.1",
    "ragas>=0.3",
    "sagemaker>=2.250,<3",
    "scikit-learn>=1.7.2",
    "sentence-transformers>=5.1.0",
    "strands-agents-tools>=0.2.8",
    "strands-agents[otel,sagemaker]>=1.10",
    "transformers>=4.56.1",
]

[dependency-groups]
dev = [
    "ruff>=0.13.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["hallucination_utils"]
